backend/requirements.txt
Django==3.0
psycopg2==2.8.5
djangorestframework==3.11.0
django-cors-headers==3.2.1
redis==3.4.1
django-redis-cache==2.1.1
celery==4.4.1
pytest-django==3.8.0
